refactor: put filter functionality inside metadata
