+++ src/func/collection.rs
@@ -636,13 +636,20 @@ impl Collection {
         Ok(ids)
     }
 
-    /// Filters the collection to get matching vector records.
-    /// * `filter`: Filter to apply to the records.
+    /// Filters the collection metadata to get matching vector records.
+    /// * `filter`: Metadata filter to apply to the records.
     pub fn filter(
         &self,
-        filter: &Filter,
+        filter: &Metadata,
     ) -> Result<HashMap<VectorID, Record>, Error> {
-        let ids = filter.get_vector_ids(&self.data);
+        // Get the IDs of the records that match the filter.
+        let data_iter = self.data.par_iter();
+        let ids: Vec<VectorID> = data_iter
+            .filter_map(|(id, data)| match data.match_filter(filter) {
+                true => Some(*id),
+                false => None,
+            })
+            .collect();
 
         // Use this with .collect() to convert the tuple to a HashMap.
         let map_result = |id: &VectorID| {
