+++ src/prelude/mod.rs
@@ -2,6 +2,5 @@ pub use crate::database::*;
 pub use crate::func::collection::*;
 pub use crate::func::distance::*;
 pub use crate::func::err::*;
-pub use crate::func::filter::*;
 pub use crate::func::metadata::*;
 pub use crate::func::vector::*;
